Check user's answer against Game
Created a new function 'checkAnswer', this is called when a user clicks
a button. The index of the last answer is passed to the function.
This user answer is then checked against the gamePattern. If it is
correct 'nextSequence' is called after a delay and the user pattern is
reset.